+++ Models/GolfboxMarkerModels.cs
@@ -1,0 +1,9 @@
+namespace GolfkollektivetBackend.Models;
+
+public class MarkerSearchResult
+{
+    public string Guid { get; set; } = default!;
+    public string Name { get; set; } = default!;
+    public string Club { get; set; } = default!;
+    public string Display { get; set; } = default!;
+}
