+++ Models/TeeOption.cs
@@ -1,0 +1,6 @@
+public class TeeOption
+{
+    public string Name { get; set; } = default!;
+    public string Gender { get; set; } = default!;
+    public string Guid { get; set; } = default!;
+}
